Remove the `no-loop-func` rule
It has too many false positives and I can’t say I’ve ever been hit or seen the mistakes it’s supposed to prevent. It’s pretty much just an annoyance at this point.

For example, it reports about doing a `.map()` iteration inside a for-of loop, which is totally valid.